fix: migration idempotency for remote DB execution
- 008: replace gen_random_bytes with gen_random_uuid (pgcrypto schema issue)
- 014: wrap policies in DO block, handle existing notifications table
- 017: change feed_score params from INTEGER to BIGINT (COUNT returns BIGINT)
- 027-030: add DROP POLICY IF EXISTS before CREATE to prevent conflicts

+++ supabase/migrations/017_feed_algorithm.sql
@@ -13,10 +13,10 @@ CREATE OR REPLACE FUNCTION feed_score(
   p_likes INTEGER,
   p_replies INTEGER,
   p_created_at TIMESTAMPTZ,
-  p_reaction_fire INTEGER DEFAULT 0,
-  p_reaction_game INTEGER DEFAULT 0,
-  p_reaction_art INTEGER DEFAULT 0,
-  p_reaction_mindblown INTEGER DEFAULT 0
+  p_reaction_fire BIGINT DEFAULT 0,
+  p_reaction_game BIGINT DEFAULT 0,
+  p_reaction_art BIGINT DEFAULT 0,
+  p_reaction_mindblown BIGINT DEFAULT 0
 )
 RETURNS FLOAT
 LANGUAGE plpgsql
